Split files into implementation and test modules (#5)
* Split write.rs and source.rs

* made generics more consistent
